Add YK5NFC and SKY3 to yubico-metadata.json
The extension selector for the YubiKey 5 series is changed from matching
the extension `1.3.6.1.4.1.41482.1.7` (Yubico product identifier) to
instead matching the extension `1.3.6.1.4.1.45724.1.1.4`
(`id-fido-gen-ce-aaguid`). This is because all YubiKey 5 series devices
have the same value for the `1.3.6.1.4.1.41482.1.7` extension, while the
YubiKey 5 NFC has a different AAGUID (`1.3.6.1.4.1.45724.1.1.4`) than
the others.

This also introduces a new value format along with the new AAGUID
extension. This won't work with pre-existing reader implementations.