Expand the YubiHSM code signing example
Updated the provided .inf file in the signing example to include some
additional properties. Specifically, the
`KeyUsageProperty = NCRYPT_ALLOW_SIGNING_FLAG`
property removes the need for the target authentication key to have the
`decrypt-pkcs` delegated capability.

The RequestType was changed from CMC to PKCS10 since that's the more
common request format. HashAlgorithm and KeyAlgorithm were added to make those
values explicit and to prevent mistaken usage of sha1. Some minor typos were
also fixed.

The code signing example was expanded to include a section on configuring
an authentication key to be used with the KSP. Of specific note are the
capabilities that are required when generating a CSR through the KSP.

Closes #252