Use mod_rewrite to add trailing slash for dirs
This is needed since mod_rewrite is executed before mod_dir

If we leave this to mod_dir's DirectorySlash-directive (default on) the
redirect will use the protocol the request came in on, which will be
http since we terminate SSL in the layer above.

By doing the redirect with a RewriteRule we can include our HTTP/HTTPS
handling logic in these redirects aswell.