Link to more relevant PIV guide from main SSH page
From https://developers.yubico.com/SSH/ it really makes more sense to
link to
https://developers.yubico.com/PIV/Guides/SSH_with_PIV_and_PKCS11.html,
in it being more in line with what the GPG and FIDO guides provide.

SSH certificates are great, but might not necessarily be the most
obvious starting point for someone who simply wants to start using
their YubiKey for SSH logins. Especially not following the
https://developers.yubico.com/PIV/Guides/SSH_user_certificates.html
guide, which risks leaving someone with an SSH CA residing on local
disk, nullifying the whole security benefit of having the SSH key on
the YubiKey.